components/auth_service/requirements.txt
fastapi==0.85.1
uvicorn==0.19.0
firebase-admin==4.5.2
redis==4.5.4
google-cloud-logging==3.5.0
protobuf==3.20.3
fireo==1.6.2
requests==2.31.0
regex==2023.6.3
